requirements.txt
yfinance>=0.2.40
ta>=0.11.0
pandas>=2.2.0
numpy>=1.26.0
firebase-admin>=6.5.0
httpx>=0.27.0
python-dotenv>=1.0.0
